requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0

# Database
SQLAlchemy>=2.0.35
psycopg2-binary==2.9.9

# Background Tasks
APScheduler==3.10.4

# AI & External APIs
openai==1.3.0

# Security & Authentication
Werkzeug==3.0.1
WTForms==3.1.1
email-validator==2.1.0

# Configuration
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# Monitoring & Logging
python-json-logger==2.0.7
